requirements.txt
PyGithub==1.59.1
click==8.1.8
requests==2.31.0
python-dotenv==1.0.0
colorama==0.4.6
tabulate==0.9.0
tqdm==4.66.4
